Use JWT to authenticate
