+++ backend/auth/utils.py
@@ -1,0 +1,26 @@
+from uuid import UUID
+from rest_framework_simplejwt.authentication import JWTAuthentication
+JWT_authenticator = JWTAuthentication()
+
+def isUUID(val):
+    try:
+        UUID(val)
+        return True
+    except:
+        return False
+
+def isAuthorized(request, author_uuid):
+    ''' Checks if the requester is authorized to do whatever method that's requested.
+        author_uuid = uuid of the author in the requested url
+    '''
+    res = JWT_authenticator.authenticate(request)
+    if res is not None:
+        user, token = res;
+        requesterUUID = user.uuid
+
+        # if the requester is not what they say they are (aren't the actual author)
+        if str(requesterUUID) != author_uuid:
+            return False
+        return True
+    else:
+        return False
