+++ backend/author/serializers.py
@@ -5,7 +5,7 @@ class AuthorSerializer(serializers.ModelSerializer):
     class Meta:
         model = Author
         fields = ['id','uuid','host','displayName','url',
-        'github','profileImage','is_active','is_superuser', 'type']
+        'github','profileImage','is_active','is_superuser', 'type', 'followers']
         extra_kwargs = {
             'type': {'read_only': True},
             'id': {'read_only': True},
