--- backend/author/viewsets.py
@@ -1,26 +1,0 @@
-from serializers import AuthorSerializer
-from models import Author
-from rest_framework import viewsets
-from rest_framework.permissions import IsAuthenticated
-from rest_framework import filters
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    http_method_names = ['get']
-    serializer_class = AuthorSerializer
-    permission_classes = (IsAuthenticated,)
-    filter_backends = [filters.OrderingFilter]
-    ordering_fields = ['updated']
-    ordering = ['-updated']
-
-    def get_queryset(self):
-        if self.request.user.is_superuser:
-            return Author.objects.all()
-
-    def get_object(self):
-        lookup_field_value = self.kwargs[self.lookup_field]
-
-        obj = Author.objects.get(lookup_field_value)
-        self.check_object_permissions(self.request, obj)
-
-        return obj
